+++ frontend/src/components/ui/StationListWindow/EmptyStationsNotice.tsx
