+++ frontend/src/components/ui/StationListWindow/StationSummaryCardSkeleton.stories.tsx
