+++ frontend/src/components/ui/StationListWindow/StationSummaryCardSkeleton.tsx
