+++ frontend/src/components/ui/StationListWindow/hooks/useFetchStationSummaries.ts
